pyproject.toml
[project]
name = "pylocksmith"
version = "0.1.0"
description = "Security Libraries We Love"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "argon2-cffi>=25.1.0",
    "bandit>=1.8.6",
    "bleach>=6.3.0",
    "cryptography>=46.0.3",
    "gunicorn>=23.0.0",
    "hashids>=1.3.1",
    "icecream>=2.1.8",
    "itsdangerous>=2.2.0",
    "pyjwt>=2.9.0",
    "passlib>=1.7.4",
    "pycryptodome>=3.23.0",
    "pynacl>=1.6.0",
    "python-decouple>=3.8",
    "python-gnupg>=0.5.5",
    "secure>=0.3.0",
    "pyopenssl>=25.3.0",
    "hvac>=2.4.0",
    "python-dotenv>=1.2.1",
    "bcrypt>=5.0.0",
]
